nixos-rebuild-ng: use namespaced imports for services
