+++ pkgs/by-name/ni/nixos-rebuild-ng/src/tests/test_main.py
@@ -8,7 +8,6 @@ from typing import Any
 from unittest.mock import ANY, Mock, call, patch
 
 import pytest
-from pytest import MonkeyPatch
 
 import nixos_rebuild as nr
 from nixos_rebuild.constants import WITH_NIX_2_18
@@ -126,92 +125,6 @@ def test_parse_args() -> None:
         "--include",
         "include2",
     ]
-
-
-@patch.dict(os.environ, {}, clear=True)
-@patch("os.execve", autospec=True)
-@patch(get_qualified_name(nr.nix.build), autospec=True)
-def test_reexec(mock_build: Mock, mock_execve: Mock, monkeypatch: MonkeyPatch) -> None:
-    monkeypatch.setattr(nr.services, "EXECUTABLE", "nixos-rebuild-ng")
-    argv = ["/path/bin/nixos-rebuild-ng", "switch", "--no-flake"]
-    args, _ = nr.parse_args(argv)
-    mock_build.return_value = Path("/path")
-
-    nr.reexec(argv, args, {"build": True}, {"flake": True})
-    mock_build.assert_has_calls(
-        [
-            call(
-                nr.services.NIXOS_REBUILD_ATTR,
-                nr.models.BuildAttr(ANY, ANY),
-                {"build": True, "no_out_link": True},
-            )
-        ]
-    )
-    # do not exec if there is no new version
-    mock_execve.assert_not_called()
-
-    mock_build.return_value = Path("/path/new")
-
-    nr.reexec(argv, args, {}, {})
-    # exec in the new version successfully
-    mock_execve.assert_called_once_with(
-        Path("/path/new/bin/nixos-rebuild-ng"),
-        ["/path/bin/nixos-rebuild-ng", "switch", "--no-flake"],
-        {"_NIXOS_REBUILD_REEXEC": "1"},
-    )
-
-    mock_execve.reset_mock()
-    mock_execve.side_effect = [OSError("BOOM"), None]
-
-    nr.reexec(argv, args, {}, {})
-    # exec in the previous version if the new version fails
-    mock_execve.assert_any_call(
-        Path("/path/bin/nixos-rebuild-ng"),
-        ["/path/bin/nixos-rebuild-ng", "switch", "--no-flake"],
-        {"_NIXOS_REBUILD_REEXEC": "1"},
-    )
-
-
-@patch.dict(os.environ, {}, clear=True)
-@patch("os.execve", autospec=True)
-@patch(get_qualified_name(nr.nix.build_flake), autospec=True)
-def test_reexec_flake(
-    mock_build: Mock, mock_execve: Mock, monkeypatch: MonkeyPatch
-) -> None:
-    monkeypatch.setattr(nr.services, "EXECUTABLE", "nixos-rebuild-ng")
-    argv = ["/path/bin/nixos-rebuild-ng", "switch", "--flake"]
-    args, _ = nr.parse_args(argv)
-    mock_build.return_value = Path("/path")
-
-    nr.reexec(argv, args, {"build": True}, {"flake": True})
-    mock_build.assert_called_once_with(
-        nr.services.NIXOS_REBUILD_ATTR,
-        nr.models.Flake(ANY, ANY),
-        {"flake": True, "no_link": True},
-    )
-    # do not exec if there is no new version
-    mock_execve.assert_not_called()
-
-    mock_build.return_value = Path("/path/new")
-
-    nr.reexec(argv, args, {}, {})
-    # exec in the new version successfully
-    mock_execve.assert_called_once_with(
-        Path("/path/new/bin/nixos-rebuild-ng"),
-        ["/path/bin/nixos-rebuild-ng", "switch", "--flake"],
-        {"_NIXOS_REBUILD_REEXEC": "1"},
-    )
-
-    mock_execve.reset_mock()
-    mock_execve.side_effect = [OSError("BOOM"), None]
-
-    nr.reexec(argv, args, {}, {})
-    # exec in the previous version if the new version fails
-    mock_execve.assert_any_call(
-        Path("/path/bin/nixos-rebuild-ng"),
-        ["/path/bin/nixos-rebuild-ng", "switch", "--flake"],
-        {"_NIXOS_REBUILD_REEXEC": "1"},
-    )
 
 
 @patch.dict(
